prev

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -8,7 +8,7 @@
 int _atoi(char *s)
 {
 	int i = 0, sign = 1;
-	long int num = 0;
+	int num = 0;
 
 	while (s[i] != '\0')
 	{
